Integrate other smoke tests into "SmokeTest"
Separate test suites "NamespaceSmokeTest" and "StubsSmokeTest"
were created when there was no support for supplying
different "command-line" options to different smoke tests. Now
this is possible through an additional text file. So
merged them into "SmokeTest".

With SmokeTest becoming the only child class of
AcceptanceTestBase, merged these two classes as well.

Resolves: #725
