+++ gluecodium/src/test/resources/smoke/namespace_basic/input/Basic.lime
