+++ gluecodium/src/test/resources/smoke/namespace_basic/input/commandlineoptions.txt
@@ -1,0 +1,7 @@
+-input $INPUT_FOLDER
+-auxinput $AUX_FOLDER
+-intnamespace gluecodium
+-javanonnullannotation android.support.annotation.NonNull
+-javanullableannotation android.support.annotation.Nullable
+-cppnamespace root.space
+-intpackage foo.bar
