+++ gluecodium/src/test/resources/smoke/namespace_basic/output/android/com/example/foo/bar/NativeBase.java
