+++ gluecodium/src/test/resources/smoke/namespace_basic/output/android/jni/com_example_smoke_SomeStruct__Conversion.cpp
