+++ gluecodium/src/test/resources/smoke/namespace_basic/output/cbridge/src/root/space/smoke/cbridge_BasicTypes.cpp
