+++ gluecodium/src/test/resources/smoke/namespace_basic/output/cpp/include/root/space/smoke/BasicTypes.h
