+++ gluecodium/src/test/resources/smoke/namespace_basic/output/lime/smoke/Basic.lime
