+++ gluecodium/src/test/resources/smoke/stubs_classes/input/Classes.lime
