+++ gluecodium/src/test/resources/smoke/stubs_structs/input/Structs.lime
