+++ gluecodium/src/test/resources/smoke/stubs_structs/input/commandlineoptions.txt
@@ -1,0 +1,6 @@
+-input $INPUT_FOLDER
+-auxinput $AUX_FOLDER
+-intnamespace gluecodium
+-javanonnullannotation android.support.annotation.NonNull
+-javanullableannotation android.support.annotation.Nullable
+-stubs true
